Create the concept of a data schema. Refactor to use data schema.
