+++ src/services/tools/query/base.ts
@@ -1,43 +1,27 @@
 import { Tool } from "@langchain/core/tools";
 import { IContext } from "@verida/types";
-import { z } from "zod";
-
-const CouchDBQuerySchema = z.object({
-    selector: z.record(z.string(), z.any()).describe(`Couchdb selector object`),
-    fields: z.array(z.string()).optional().describe(`Array of fields to include in the result`),
-    sort: z
-      .array(
-        z.record(z.string(), z.enum(["asc", "desc"])).describe(`Example: [{ "sentAt": "desc" }]`),
-      )
-      .optional(),
-    limit: z.number().optional().describe(`Maximum number of documents to return`),
-    skip: z.number().optional().describe(`Number of documents to skip`),
-    count: z.boolean().optional().describe(`If true, a count of the number of matching results will be returned`)
-});
-
-type CouchDBQuerySchemaType = z.infer<typeof CouchDBQuerySchema>;
-
-export interface BaseQueryToolConfig {
-    schemaDefinition: string
-    schemaUrl: string
-    defaultParams: Partial<CouchDBQuerySchemaType>
-    extraDetail: string
-}
-
-// ['name', 'type', 'fromName', 'fromEmail', 'messageText', 'sentAt']
+import { BaseQueryToolConfig, CouchDBQuerySchemaType } from "../../../services/interfaces";
 
 export class BaseQueryTool extends Tool {
   private context: IContext
+  private config: BaseQueryToolConfig
 
-  constructor(context: IContext) {
+  name = ""
+  description = ""
+
+  constructor(context: IContext, config: BaseQueryToolConfig) {
     super()
     this.context = context
+    this.config = config
+    this.description = `Input to this tool is a detailed and correct CouchDB query of ${this.getConfig().extraDetail} with the following schema ${this.getConfig().schemaDefinition}.
+    Set count=true to return the number of matching results, rather than the results themselves.
+    If the query is not correct, an error message will be returned.
+    If an error is returned, rewrite the query, check the query, and try again.
+    The default result limit is ${this.getConfig().defaultParams.limit}`
   }
 
-  name = "Unknown"
-
-  protected getConfig(): BaseQueryToolConfig {
-    throw new Error(`getConfig() not implemented`)
+  public getConfig(): BaseQueryToolConfig {
+    return this.config
   }
 
   /** @ignore */
@@ -89,11 +73,5 @@ export class BaseQueryTool extends Tool {
       return `${error}`;
     }
   }
-
-  description = `Input to this tool is a detailed and correct CouchDB query of ${this.getConfig().extraDetail} with the following schema ${this.getConfig().schemaDefinition}.
-  Set count=true to return the number of matching results, rather than the results themselves.
-  If the query is not correct, an error message will be returned.
-  If an error is returned, rewrite the query, check the query, and try again.
-  The default result limit is ${this.getConfig().defaultParams.limit}`
 
 }
